Extract getFunctionName to utils.ts and use it in kernel
