+++ test/utils/utils.test.ts
@@ -1,0 +1,24 @@
+import { expect } from "chai";
+import { getFunctionName } from "../../src/utils/utils";
+
+describe("Utils", () => {
+    it("Should return a good function name", () => {
+        function testFunction() {
+            return false;
+        }
+
+        expect(getFunctionName(testFunction)).eql("testFunction");
+    });
+
+    it("Should return a good function name by using the regex", () => {
+        const testFunction: any = { name: null };
+        testFunction.toString = () => {
+            return "function testFunction";
+        };
+
+        console.log(getFunctionName(testFunction));
+
+        expect(getFunctionName(testFunction)).eql("testFunction");
+    });
+
+});
